fix: resolve cyclic-import alerts by lifting; drop re-export shims; cleanups
Properly resolve the CodeQL py/cyclic-import alerts by lifting shared code
into leaf modules instead of deferring imports into function bodies, so all
imports stay at module top (TYPE_CHECKING aside):

- compilation_unit <-> core_analyses <-> interprocedural: extract
  extract_class_names() into compiler/class_names.py; hoist the analysis
  imports to module level.
- green_tree <-> lexer <-> token_scanning <-> registry.runtime: extract the
  active-dialect ContextVar into compiler/dialect_context.py and move
  is_simple_scalar_var_word into token_scanning; hoist the down-edges to top.
- explorer cli <-> tui: lift the shared rendering layer into
  tooling/explorer/_render.py.

Forward-only cleanup (no stable external interfaces): eliminate re-export
shims, repointing consumers to each symbol's real home --- delete
compiler/parsing/command_shapes.py, drop the token_helpers / runtime
(_dialect_var) / workspace_file_ops (path_to_uri) re-exports, and remove the
now-dead import-linter ignore. Remove the dead backward-compat lexer class
attributes (strict_quoting, irules_brace_separator), now thread-local.

Replace py/unnecessary-lambda (and other non-clarifying lambdas) with
operator.* / functools.partial / named callables; keep idiomatic sort-key and
dispatch lambdas.

Verified: ty + ruff + import-linter clean; cyclic-import confirmed gone via a
points-to-faithful detector and CodeQL; uninitialized-local fix confirmed by a
focused CodeQL run; full test-slow phases pass.
