+++ compiler/interprocedural.py
@@ -18,6 +18,8 @@ from typing import TYPE_CHECKING
 
 from compiler.analysis_types import LatticeKind, LatticeValue
 from compiler.cfg import CFGFunction, CFGReturn, build_cfg
+from compiler.core_analyses import _expr_has_command as _expr_has_command_sub
+from compiler.core_analyses import analyse_function
 from compiler.eval_helpers import DECIMAL_INT_RE as _DECIMAL_INT_RE
 from compiler.expr_ast import (
     ExprBinary,
@@ -51,7 +53,7 @@ from compiler.ir import (
 )
 from compiler.lowering import lower_to_ir
 from compiler.parsing.green_tree import tokenise
-from compiler.parsing.lexer import is_simple_scalar_var_word
+from compiler.parsing.token_scanning import is_simple_scalar_var_word
 from compiler.proc_arg_traits import (
     infer_param_traits,
     infer_param_traits_deep,
@@ -583,10 +585,6 @@ def _scan_local_facts(
     known_procs: set[str],
     facts: _LocalFacts,
 ) -> None:
-    from compiler.core_analyses import (
-        _expr_has_command as _expr_has_command_sub,  # deferred (cycle)
-    )
-
     for stmt in script.statements:
         if isinstance(stmt, IRBarrier):
             facts.has_barrier = True
@@ -1166,8 +1164,6 @@ def analyse_interprocedural_ir(
     facts are reused for unchanged procedures using key
     ``(qualified_name, hash(proc_source_text))``.
     """
-    from compiler.core_analyses import analyse_function  # deferred (cycle)
-
     if not ir_module.procedures and not ir_module.methods:
         return InterproceduralAnalysis(procedures={})
 
@@ -1543,8 +1539,6 @@ def evaluate_proc_with_constants(
 
     Returns the constant return value if determinable, else None.
     """
-    from compiler.core_analyses import analyse_function  # deferred (cycle)
-
     if len(args) != len(params):
         return None
     param_constants = {(p, 0): LatticeValue.const(a) for p, a in zip(params, args)}
