+++ server/features/workspace_file_ops.py
@@ -18,7 +18,7 @@ from analyser.semantic_model import AnalysisResult, SourceTarget
 from analyser.source_resolver import resolve_source_target
 from server._lsp_conv import to_lsp_range
 
-from ..workspace.scanner import path_to_uri, uri_to_path  # noqa: F401  (path_to_uri re-exported)
+from ..workspace.scanner import uri_to_path
 
 
 class _WorkspaceLike(Protocol):
