+++ tests/test_parsing_helpers.py
@@ -3,9 +3,9 @@
 from __future__ import annotations
 
 from compiler.parsing.argv import widen_argv_tokens_to_word_spans
-from compiler.parsing.command_shapes import extract_single_expr_argument
 from compiler.parsing.known_commands import known_command_names
 from compiler.parsing.lexer import TclLexer
+from compiler.parsing.token_scanning import extract_single_expr_argument
 from compiler.registry import REGISTRY
 from shared.tokens import Token, TokenType
