+++ tooling/tcl/verbs/_registry.py
@@ -8,7 +8,7 @@ every verb to an argparse subparser group in a single call.
 from __future__ import annotations
 
 import argparse
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import Callable
 
 
@@ -18,7 +18,7 @@ class _VerbSpec:
     configure: Callable
     aliases: tuple[str, ...]
     help: str
-    formatter_class: type = field(default_factory=lambda: argparse.HelpFormatter)
+    formatter_class: type = argparse.HelpFormatter
 
 
 _VERB_REGISTRY: list[_VerbSpec] = []
